fix: unit-return handling, builtin-name collection, and Go import gaps
- required_includes: add missing entry for `open` (C: fcntl.h,
  sys/stat.h, string.h; C++: fcntl.h, sys/stat.h, string) — verified
  against -Wall -Werror -Wextra -std=c11/c++17

- BulletBody::Builtin(name) shorthand now respects returns_unit in
  codegen_c.rs, codegen_cpp.rs, and codegen.rs (Rust). Rust case is a
  distinct failure mode from C/C++: a unit-typed function previously
  placed the builtin call in tail-expression position, which type-
  mismatches when the builtin's Rust code evaluates to a non-unit
  value (e.g. close()'s `{ ...; 0i32 }`). Exposed output_is_unit/
  type_is_unit as pub(crate) in codegen_c.rs for reuse.

- collect_builtin_names_c (C/C++, shared) and collect_builtin_names
  (Rust) only ever scanned BulletBody::Pipes, never the bare
  BulletBody::Builtin(name) shorthand — silently dropping required
  #includes/use lines, hoisted static helpers, and foreign_types.h
  detection for any function using the shorthand form. Fixed in all
  three call sites.

- Go: builtin_imports() had no entries for open/close/out/in/env/
  exit/time/sleep/run/args — generated Go code using any of these
  failed with `undefined: os` etc. Added the missing os/syscall/
  time/bufio/runtime/os-exec entries.

- Go: close()/fd_out() emitted syscall.Close(uintptr(fd))/
  syscall.Write(uintptr(fd), ...) — wrong argument type, both take
  int on this platform. Fixed to int(fd).

Verified for real: gcc/g++ against the project's mandatory flags plus
personal convention flags; go build + run for all 9 previously-broken
Go builtins (installed golang-go via apt for this). Rust-side changes
are stub-validated only — no cargo build available in this sandbox
(same edition2024/cargo 1.75 limitation as last session) — real
cargo build still needed on your end.

+++ src/stdlib/close.rs
@@ -46,7 +46,7 @@ pub fn emit(params: &[Param], backend: &Backend) -> Result<String, String> {
         // syscall.Close returns an error; normalise to 0 / -1.
         Backend::Go => format!(
             "func() int32 {{\
-               if __err := syscall.Close(uintptr({fd})); __err != nil {{ return -1 }}\
+               if __err := syscall.Close(int({fd})); __err != nil {{ return -1 }}\
                return 0\
              }}()"
         ),
